[IMP] nortenas: fill salesman when creating credit note
When a credit note is created from an invoice, the salesman field
(salesperson_id, not the native one) is filled automatically. since the
salesman field was created using Odoo Studio, the following is
performed:
- Rename/deactivate old fields
- Create them on this module
- Create on this module records which depends on them when possible
  (views and reports) and update directly when not (automated actions)

In addition, the salesman field may now be edited when a invoice/credit
note is created from scratch, i.e. not from a sale order/invoice.