[FIX] nortenas: External ID of operation type on migration script
The change that introduced a new wizard to autopackage products,
modified the external ID of the operation tipe "Release product",
through a migration script, so the record was associated to the
`nortenas` module. However, the original external ID didn't exist in
production yet, so it was generated and this change updates its
reference.