[FIX] nortenas: Non iterable NewID on domain at _compute_sales_value
When we tried to create new locations it was giving an error on
`_compute_sales_value` calculating the domain elements. It was expecting
a value or a list of values but was getting a NewId object instead.

To prevent that to happen, we changed `record.id` by `record.ids` ir
order to get the right values for the domain to be computed.